Add transactions to portfolio and portfolio charts

+++ PriceFlowApp/DTOs/Transaction.cs
@@ -6,7 +6,7 @@
         public int HVId { get; set; }
         public string HVCode { get; set; } = null!;
         public int SharesQuantity { get; set; }
-        public decimal SharesUnitPrice { get; set; }
+        public int SharesUnitPrice { get; set; }
         public decimal Amount {  get; set; }
         public string TypeTransaction { get; set; } = null!;
         public bool IsReal { get; set; }
